swarm: initial implementation
- load conversations
- add conversation
- send messages
- load messages
- receive message
- remove conversation

This patch does not add support for:
- remove message
- resend messageq
- last displayed message
- clear all history
- remove contact

+++ src/callbackshandler.h
@@ -181,15 +181,15 @@ Q_SIGNALS:
                                      const QString& to,
                                      int status);
 
-    void transferStatusCreated(long long dringId, api::datatransfer::Info info);
-    void transferStatusCanceled(long long dringId, api::datatransfer::Info info);
-    void transferStatusAwaitingPeer(long long dringId, api::datatransfer::Info info);
-    void transferStatusAwaitingHost(long long dringId, api::datatransfer::Info info);
-    void transferStatusOngoing(long long dringId, api::datatransfer::Info info);
-    void transferStatusFinished(long long dringId, api::datatransfer::Info info);
-    void transferStatusError(long long dringId, api::datatransfer::Info info);
-    void transferStatusTimeoutExpired(long long dringId, api::datatransfer::Info info);
-    void transferStatusUnjoinable(long long dringId, api::datatransfer::Info info);
+    void transferStatusCreated(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusCanceled(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusAwaitingPeer(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusAwaitingHost(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusOngoing(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusFinished(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusError(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusTimeoutExpired(qulonglong dringId, api::datatransfer::Info info);
+    void transferStatusUnjoinable(qulonglong dringId, api::datatransfer::Info info);
 
     /**
      * Connect this signal to get when a device name changed or a device is added
